requirements.txt
Django==4.1
psycopg2==2.9.3
djangorestframework==3.13.1
drf_yasg==1.21.3
